requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
pydantic
pandas>=2.1.0
numpy>=1.26.0
xgboost>=2.0.0
shap>=0.45.0
scikit-learn>=1.4.0
PyMuPDF>=1.24.0
python-docx>=1.1.0
jinja2>=3.1.0
google-search-results>=2.4.0
databricks-sql-connector>=3.0.0
matplotlib>=3.9.0
